acceptance_tests: fix invalid cleanup step
as we validate all pipelines in reference_dataset/pipelines, we need to cleanup the directory before running tests. We were referencing an invalid directory, so cleanup was skipped.

+++ lib/coa/reference_dataset_documentation/docs_config.rb
@@ -45,12 +45,14 @@ module Coa
       end
 
       def cleanup_generated_pipelines
+        puts "deleting #{generated_pipeline_list}";
         FileUtils.rm_rf(generated_pipeline_list)
       end
 
       def self.cleanup_generated_pipelines(dir)
         pattern = "#{dir}/pipelines/*-generated.yml"
         generated_pipeline_list = Dir.glob(pattern)
+        raise Errno::ENOENT, "#{dir}" unless Dir.exist?(dir)
         FileUtils.rm_rf(generated_pipeline_list)
       end
     end
